requirements.txt
PyGithub==2.5.0
keyring==25.6.0
python-dateutil==2.9.0